+++ lib/ui/profile/ProfileInfoPage.dart
@@ -31,11 +31,10 @@ class ProfileInfoPage extends StatefulWidget {
 
   final ProfileInfo contentType;
   final Map<String, dynamic>? params;
-  final bool showProfileCommands;
-  final bool showAccountCommands;
+  final bool onboarding;
   final void Function()? onStateChanged;
 
-  ProfileInfoPage({super.key, required this.contentType, this.params, this.showProfileCommands = true, this.showAccountCommands = false, this.onStateChanged});
+  ProfileInfoPage({super.key, required this.contentType, this.params, this.onboarding = false, this.onStateChanged});
 
   @override
   State<StatefulWidget> createState() => ProfileInfoPageState();
@@ -63,8 +62,11 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
   bool _updatingDirectoryVisibility = false;
   bool _preparingDeleteAccount = false;
 
-  bool get isEditing => _editing;
+  bool get _showProfileCommands => (widget.onboarding == false);
+  bool get _showAccountCommands => (widget.onboarding == false);
+
   bool get isLoading => _loading;
+  bool get isEditing => _editing;
   bool get directoryVisibility => (_privacy?.public == true);
 
   void setEditing(bool value) {
@@ -115,16 +117,19 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
       return Column(children: [
         _directoryVisibilityControl,
 
-        if (directoryVisibility == true)
+        // if (directoryVisibility == true)
           Column(children: [
-            Padding(padding: EdgeInsets.symmetric(vertical: 16), child:
-              Text(_desriptionText, style: Styles().textStyles.getTextStyle('widget.detail.small'), textAlign: TextAlign.center,),
-            ),
+            if (widget.onboarding == false)
+              Padding(padding: EdgeInsets.only(top: 16), child:
+                Text(_desriptionText, style: Styles().textStyles.getTextStyle('widget.detail.small'), textAlign: TextAlign.center,),
+              ),
 
-            _editing ? _editContent : _previewContent,
+            Padding(padding: EdgeInsets.only(top: 16), child:
+              _editing ? _editContent : _previewContent,
+            )
           ]),
 
-        if (widget.showAccountCommands && !_editing)
+        if (_showAccountCommands && !_editing)
           _accountCommands,
       ],);
     }
@@ -139,11 +144,12 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
           profile: _profile,
           privacy: _privacy,
           identifiers: _identifiers,
+          onboarding: widget.onboarding,
           pronunciationAudioData: _pronunciationAudioData,
           photoImageData: _photoImageData,
           photoImageToken: _photoImageToken,
         ),
-        if (widget.showProfileCommands)
+        if (_showProfileCommands)
           Padding(padding: EdgeInsets.only(top: 24), child:
             _previewCommandBar,
           ),
@@ -157,10 +163,10 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
       profile: _profile,
       privacy: _privacy,
       identifiers: _identifiers,
+      onboarding: widget.onboarding,
       pronunciationAudioData: _pronunciationAudioData,
       photoImageData: _photoImageData,
       photoImageToken: _photoImageToken,
-      showProfileCommands: widget.showProfileCommands,
       onFinishEdit: _onFinishEditInfo,
   );
 
@@ -170,7 +176,7 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
         Row(children: [
           Expanded(child:
             Padding(padding: EdgeInsets.only(left: 16, top: 12), child:
-              Text(Localization().getStringEx('panel.profile.info.directory_visibility.command.toggle.title', 'Directory Visibility'), style: Styles().textStyles.getTextStyle('widget.detail.dark.regular.fat'),)
+              Text(Localization().getStringEx('panel.profile.info.directory_visibility.command.toggle.title', 'Directory Visibility'), style: Styles().textStyles.getTextStyle('widget.detail.regular.fat'),)
             ),
           ),
           _updatingDirectoryVisibility ? _directoryVisibilityProgress : _directoryVisibilityToggleButton,
@@ -208,21 +214,25 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
       Localization().getStringEx('panel.profile.info.directory_visibility.public.text', 'Public') :
       Localization().getStringEx('panel.profile.info.directory_visibility.private.text', 'Private');
 
-    final String messageTemplate = directoryVisibility ?
-      Localization().getStringEx('panel.profile.info.directory_visibility.public.description', 'Your directory visibility is set to $visibilityMacro. Anyone on or off the User Directory can view your account.') :
-      Localization().getStringEx('panel.profile.info.directory_visibility.private.description', 'Your directory visibility is set to $visibilityMacro. Your account is available only to you.');
+    final String messageTemplate = widget.onboarding ?
+      (directoryVisibility ?
+        AppTextUtils.appTitleString('panel.profile.info.directory_visibility.onboarding.public.description', 'Your directory visibility is set to $visibilityMacro. The information below will be visible to the users of the ${AppTextUtils.appTitleMacro} App.') :
+        AppTextUtils.appTitleString('panel.profile.info.directory_visibility.onboarding.private.description', 'Your directory visibility is set to $visibilityMacro. Your profile is visible only to you.')) :
+      (directoryVisibility ?
+        AppTextUtils.appTitleString('panel.profile.info.directory_visibility.normal.public.description', 'Your directory visibility is set to $visibilityMacro. Anyone on or off the User Directory can view your account.') :
+        AppTextUtils.appTitleString('panel.profile.info.directory_visibility.normal.private.description', 'Your directory visibility is set to $visibilityMacro. Your account is available only to you.'));
 
     final List<String> messages = messageTemplate.split(visibilityMacro);
     List<InlineSpan> spanList = <InlineSpan>[];
     if (0 < messages.length)
       spanList.add(TextSpan(text: messages.first));
     for (int index = 1; index < messages.length; index++) {
-      spanList.add(TextSpan(text: visibilityValue, style : Styles().textStyles.getTextStyle("widget.detail.dark.small.fat"),));
+      spanList.add(TextSpan(text: visibilityValue, style : Styles().textStyles.getTextStyle("widget.detail.small.fat"),));
       spanList.add(TextSpan(text: messages[index]));
     }
 
     return RichText(textAlign: TextAlign.left, text:
-      TextSpan(style: Styles().textStyles.getTextStyle("widget.detail.dark.small"), children: spanList)
+      TextSpan(style: Styles().textStyles.getTextStyle("widget.detail.small"), children: spanList)
     );
   }
 
@@ -298,13 +308,15 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
 
   Widget get _editInfoButton => RoundedButton(
     label: _editInfoButtonTitle,
-    fontFamily: Styles().fontFamilies.bold, fontSize: 16,
+    textStyle: Styles().textStyles.getTextStyle('widget.button.light.title.medium'),
+    backgroundColor: Styles().colors.background,
+    borderColor: Styles().colors.fillColorSecondary,
     padding: EdgeInsets.symmetric(horizontal: 12, vertical: 8),
     onTap: _onEditInfo,
   );
 
   String get _editInfoButtonTitle =>
-    Localization().getStringEx('panel.profile.info.command.button.edit.text', 'Edit My Info');
+    Localization().getStringEx('panel.profile.info.command.button.edit.text', 'Edit my info');
 
   Widget get _swapInfoButton => RoundedButton(
     label: Localization().getStringEx('panel.profile.info.command.button.swap.text', 'Swap Info'),
@@ -513,7 +525,7 @@ class ProfileInfoPageState extends ProfileDirectoryMyInfoBasePageState<ProfileIn
         _pronunciationAudioData = pronunciationAudioData;
       }
 
-      if (widget.showProfileCommands) {
+      if (_showProfileCommands) {
         _editing = false;
       }
       widget.onStateChanged?.call();
@@ -529,7 +541,7 @@ class ProfileDirectoryMyInfoBasePageState<T extends StatefulWidget> extends Stat
   // Name Text Style
 
   TextStyle? get nameTextStyle =>
-    Styles().textStyles.getTextStyleEx('widget.message.medium_large.fat', fontHeight: 0.85, textOverflow: TextOverflow.ellipsis);
+    Styles().textStyles.getTextStyleEx('widget.title.medium_large.fat', fontHeight: 0.85, textOverflow: TextOverflow.ellipsis);
 
   // Positive and Permitted visibility